commit changes for DB jpa

+++ manager/src/main/java/com/analysis/manager/controllers/AnimalsController.java
@@ -1,7 +1,13 @@
 package com.analysis.manager.controllers;
 
+import com.analysis.manager.Service.AnimalsService;
+import com.analysis.manager.Service.LayerService;
+import com.analysis.manager.Service.ProjectService;
+import com.analysis.manager.Service.UserService;
 import com.analysis.manager.modle.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,13 +19,16 @@ import org.springframework.web.bind.annotation.RequestParam;
 public class AnimalsController {
 
     @Autowired
-    private ProjectDao projectDao;
+    private ProjectService projectDao;
 
     @Autowired
-    private LayerDao layerDao;
+    private LayerService layerDao;
 
     @Autowired
-    private AnimalsDao animalsDao;
+    private AnimalsService animalsDao;
+
+    @Autowired
+    private UserService userService;
 
     @RequestMapping(value = "projects/{id}/animals", method = RequestMethod.POST)
     public String createAnimalForProject(@PathVariable("id") long project_id, @RequestParam("name") String name,
@@ -27,22 +36,28 @@ public class AnimalsController {
                                          Model model)
     {
         try {
-            Project project = projectDao.getById(project_id);
-            Layer layer = layerDao.getById(layer_id);
+            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+            User user = userService.findUserByEmail(auth.getName());
+            Project project = projectDao.findByIdAndUser(project_id, user);
+            Layer layer = layerDao.findByIdAndProject(layer_id, project);
 
-            for (Object animal: project.getAnimals())
-            {
-                if (((Animal)animal).getName().equals(name) && ((Animal)animal).getLayer().getId() == layer.getId())
+            if (animalsDao.findByNameAndLayer(name, layer) == null) {
+                for (Object animal: project.getAnimals())
                 {
-                    model.addAttribute("error_massage", "Error, animal already exists in this project");
-                    return "redirect:/projects/" + project_id;
+                    if (((Animal)animal).getName().equals(name) && ((Animal)animal).getLayer().getId() == layer.getId())
+                    {
+                        model.addAttribute("error_massage", "Error, animal already exists in this project");
+                        return "redirect:/projects/" + project_id;
+                    }
                 }
-            }
 
-            Animal animal = new Animal(name, description, null, layer);
-            project.AddAnimal(animal);
-            animalsDao.create(animal);
-            projectDao.update(project);
+                Animal animal = new Animal(name, description, null, layer);
+                project.AddAnimal(animal);
+                animalsDao.save(animal);
+                projectDao.saveProject(project);
+            } else {
+                model.addAttribute("error_massage", "Animal with this name already exists");
+            }
 
             return "redirect:/projects/" + project_id;
         }
@@ -57,19 +72,26 @@ public class AnimalsController {
     public String delete(@PathVariable("id") long projectId, @PathVariable("animal_id") long animal_id, Model model)
     {
         try {
-            Animal animal = animalsDao.getById(animal_id);
+            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+            User user = userService.findUserByEmail(auth.getName());
+            Project project = projectDao.findByIdAndUser(projectId, user);
+
+            Animal animal = animalsDao.findById(animal_id);
 
             if (animal == null)
             {
-                model.addAttribute("error_massage", "Can not find animal by id = " + animal_id);
+                model.addAttribute("error_massage", "Can not find animal with id = " + animal_id);
                 return "redirect:/projects/" + projectId;
             }
-
-            Project project = projectDao.getById(projectId);
-
-            project.deleteAnimal(animal);
-            projectDao.update(project);
-            animalsDao.delete(animal);
+            else {
+                if (project.getLayers().contains(animal.getLayer())){
+                    project.deleteAnimal(animal);
+                    projectDao.saveProject(project);
+                    animalsDao.delete(animal);
+                } else {
+                    model.addAttribute("error_massage", "Can not delete animal that belong to other user");
+                }
+            }
 
             return "redirect:/projects/" + projectId;
         }
